Create .htaccess
This is to bypass extension restrictions. Even though the PHP file sends the MIME type for a PNG, it may still block the image based on the .php extension. Using this will allow you to open "domain/convert.png" with the same GET parameters and get the same results as if using "domain/index.php"